sim: Redesign the list struct to struct list_head
The old method to use the first element of the struct for the
next point. Using the new list API for the list_head struct,
the new list will default alloc a list head node. There are
simple the operation of the list.  And implement more list
on the a struct.

signed-off-by: Zhong Hongbo <bocui107@gmail.com>